--- abn/application/doc.go
@@ -1,2 +1,0 @@
-// Package application provides a runtime model of applications and their versions
-package application
