--- abn/core/doc.go
@@ -1,2 +1,0 @@
-// Package core provides a service that supports the Iter8 A/B/n SDK
-package core
