--- abn/k8sclient/doc.go
@@ -1,2 +1,0 @@
-// Package k8sclient enables interaction with Kubernetes using both a typed client and a dynamic client
-package k8sclient
